Rigaweb {diagram}	R Documentation

Gulf of Riga autumn planktonic food web

Description

Carbon flux matrix of the Gulf of Riga planktonic food web in autumn as reconstructed by inverse modelling by Donali et al. (1999).

The Gulf of Riga is a highly eutrophic system in the Baltic Sea.

The foodweb comprises 7 functional compartments:

picoautotrophs (P1)
non-picoautotrophs (P2)
bacteria (B)
heterotrophic nanoflagellates (N)
zooplankton (Z)
detritus, including virus (D)
dissolved organic carbon (DOC)

and two external compartments:

CO2
Sedimentation

These compartments are connected with 26 flows.

Units of the flows are mg C/m3/day.

Usage

Rigaweb

Format

matrix with flow values, where element ij denotes flow from compartment i to j

rownames and columnames are the components.

Author(s)

Karline Soetaert <k.soetaert@nioo.knaw.nl>.

References

Donali, E., Olli, K., Heiskanen, A.S., Andersen, T., 1999. Carbon flow patterns in the planktonic food web of the Gulf of Riga, the Baltic Sea: a reconstruction by the inverse method. Journal of Marine Systems 23, pp. 251 268.

See Also

Takapotoweb

Examples

  plotweb(Rigaweb,main="Gulf of Riga planktonic food web",
                    sub="mgC/m3/day")

[Package diagram version 1.5 Index]