Takapotoweb {diagram}	R Documentation

Takapoto atoll planktonic food web

Description

Carbon flux matrix of the Takapoto atoll planktonic food web

as reconstructed by inverse modelling by Niquil et al. (1998).

The Takapoto Atoll lagoon is located in the French Polynesia of the South Pacific

The food web comprises 7 functional compartments:

Phytoplankton
Bacteria
Protozoa
Microzooplankton
Mesozooplankton
Detritus
Dissolved organic carbon (DOC)

and three external compartments/sinks:

CO2
Sedimentation
Grazing

These compartments are connected with 32 flows. Units of the flows are mg C/m2/day

Usage

Takapotoweb

Format

matrix with flow values, where element ij denotes flow from compartment i to j

rownames and columnames are the components.

Author(s)

Karline Soetaert <k.soetaert@nioo.knaw.nl>

References

Niquil, N., Jackson, G.A., Legendre, L., Delesalle, B., 1998. Inverse model analysis of the planktonic food web of Takapoto Atoll (French Polynesia). Marine Ecology Progress Series 165, pp. 17 29.

See Also

Rigaweb

Examples

plotweb(Takapotoweb,main="Takapoto atoll planktonic food web",
                    sub="mgC/m2/day",lab.size=1)

[Package diagram version 1.5 Index]