diagram-package {diagram}	R Documentation

Functions for visualising simple graphs (networks), plotting flow diagrams

Description

Visualises simple graphs (networks) based on a transition matrix, utilities to plot flow diagrams, visualising webs,...

Support for the book "A practical guide to ecological modelling - using R as a simulation platform" by Karline Soetaert and Peter M.J. Herman (2009). Springer.

Details

Package: 	 diagram

Type: 	 Package

Version: 	 1.5

Date: 	 2009-10-10

License: 	 GNU Public License 2 or above

This package is used in R-package ecolMod, which includes many more examples.

Author(s)

Karline Soetaert (Maintainer)

See Also

plotmat, plotweb, coordinates, openplotmat, bentarrow, curvedarrow, segmentarrow, selfarrow, splitarrow, straightarrow, treearrow, shadowbox, textdiamond, textellipse, textempty, texthexa, textdiamond, textplain, textrect, textround.

Examples

## Not run: 
## show examples (see respective help pages for details)
example(plotmat)
example(plotweb)

## run demos
demo("flowchart") # creating flow charts
demo("plotmat")   # plotting diagrams inputted as a matrix
demo("plotweb")   # plotting webs inputted as a matrix

## open the directory with source code of demos
browseURL(paste(system.file(package="diagram"), "/demo", sep=""))

## show package vignette
vignette("diagram")
edit(vignette("diagram"))
browseURL(paste(system.file(package="diagram"), "/doc", sep=""))
## End(Not run)

[Package diagram version 1.5 Index]