textellipse {diagram}	R Documentation

adds lines of text in an ellipsoid box to a plot

Description

adds one or more lines of text, centered around "mid" in an ellipsoid box

Usage

textellipse(mid, radx, rady=radx*length(lab), lwd=1,  
  shadow.size=0.01, adj=c(0.5,0.5),lab="", box.col="white", 
  lcol="black", shadow.col="grey", angle=0, dr=0.01, ...)

Arguments

mid 	 midpoint (x,y) of the box.

radx 	 horizontal radius of the box.

rady 	 vertical radius of the box.

lwd 	 line width of line surrounding the box.

shadow.size 	 relative size of shadow.

adj 	 text adjustment.

lab 	 one label or a vector string of labels to be added in box.

box.col 	 fill color of the box.

lcol 	 line color surrounding box.

shadow.col 	 color of shadow.

angle 	 rotation angle, degrees.

dr 	 size of segments, in radians, to draw ellipse (decrease for smoother).

... 	 other arguments passed to function textplain.

Details

see shadowbox for specifications of the ellipsoid-shaped box and its shadow

Author(s)

Karline Soetaert <k.soetaert@nioo.knaw.nl>

See Also

textdiamond, textempty, texthexa, textmulti, textplain, textrect, textround

Examples

  openplotmat(xlim=c(-0.1,1.1),main="textellipse")
  for (i in 1:10) textellipse(mid=runif(2),col=i,box.col=grey(0.95),radx=0.1,
                          rady=0.05,lab=LETTERS[i],cex=2,angle=runif(1)*360)

[Package diagram version 1.5 Index]