textempty {diagram}	R Documentation

adds lines of text, on a colored background to a plot

Description

adds one or more lines of text, with a colored background, no box

Usage

textempty(mid, lab="", adj=c(0.5,0.5), box.col="white", cex=1, ...)

Arguments

mid 	 midpoint (x,y) of the text.

lab 	 one label or a vector string of labels to be added in box.

adj 	 text adjustment.

box.col 	 background color.

cex 	 relative size of text.

... 	 other arguments passed to function textplain.

Author(s)

Karline Soetaert <k.soetaert@nioo.knaw.nl>

See Also

textdiamond, textellipse, texthexa, textmulti, textplain, textrect, textround

Examples

  openplotmat(xlim=c(-0.1,1.1),col="lightgrey",main="textempty")
  for (i in 1:10) textempty(mid=runif(2),box.col=i,lab=LETTERS[i],cex=2)
  textempty(mid=c(0.5,0.5),adj=c(0,0),lab="textempty",box.col="white")

[Package diagram version 1.5 Index]