textplain {diagram}	R Documentation

adds lines of text to a plot

Description

adds one or more lines of text, centered around "mid"

Usage

  textplain(mid, height=0.1, lab="", adj=c(0.5,0.5), ...)

Arguments

mid 	 central coordinates where to write the text.

height 	 height of text.

lab 	 one or more character strings or expressions specifying the *text* to be written.

adj 	 label adjustments.

... 	 other arguments passed to R-function text.

Author(s)

Karline Soetaert <k.soetaert@nioo.knaw.nl>

See Also

textdiamond, textellipse, textempty, texthexa, textmulti, textrect, textround

Examples

  openplotmat(main="textplain")
  textplain(mid=c(0.5,0.5), 
            lab=c("this text is","centered", "4 strings", "on 4 lines"))
  textplain(mid=c(0.5,0.2), adj=c(0,0.5), font=2, height=0.05,
            lab=c("this text is","left alligned"))
  textplain(mid=c(0.5,0.8),adj=c(1,0.5),font=3,height=0.05, 
            lab=c("this text is","right alligned"))

[Package diagram version 1.5 Index]